discover method should return the current state of all devices on the network
With the current library there is no way to know when WeMo devices have been removed from the local network. Currently the discover method only returned newly discovered devices.

This change ensures that all devices returned represent the current state of the WeMo devices on the network. Not only is this more accurate but it allows the consumer to understand when devices have been removed from the network.